fix region bugs, change ws2812 library

+++ src/effects/spectrum_cycle.hh
@@ -10,11 +10,11 @@
 
 class SpectrumCycle : public Effect {
 public:
-    explicit SpectrumCycle(Region &region, Color base, uint16_t delay = 1);
+    explicit SpectrumCycle(Region &region, Color base, double delay = 1);
     void init() override;
     int16_t loop() override;
 private:
     const Color base;
     ColorCycling active;
-    const uint16_t delay;
+    const int16_t delay;
 };
